fix printf format problem on machines with long tv_usec; version 2.6.7

+++ time.c
@@ -80,9 +80,9 @@ timeval_str(const struct timeval *src, bool milliseconds) {
 	struct tm result, *y = gmtime_r(&t, &result);
 	dst = ret + strftime(ret, sizeof ret, "%F %T", y);
 	if (milliseconds)
-		sprintf(dst, ".%03d", src->tv_usec % 1000);
+		sprintf(dst, ".%03ld", (long)src->tv_usec % 1000);
 	else
-		sprintf(dst, ".%06d", src->tv_usec % 1000000);
+		sprintf(dst, ".%06ld", (long)src->tv_usec % 1000000);
 	return ret;
 }
